+++ cmake/Geant4.cmake
@@ -72,6 +72,12 @@ if (Geant4_FOUND)
       # Mark Geant4_DIR as advanced since can't be set in GUI and this turned out to be confusing
       mark_as_advanced(Geant4_DIR)
       mark_as_advanced(Geant4_PREFIX)
+
+      option(USE_CUSTOM_CHANNELLING "Use custom build of crystal channelling in Geant4" OFF)
+      if(USE_CUSTOM_CHANNELLING)
+        add_definitions("-DCUSTOMCHANNELING")
+        set(PREPROCESSOR_DEFS "-DCUSTOMCHANNELING")
+      endif()
 else(Geant4_FOUND)
       message(FATAL_ERROR "Geant4 not found. Exiting. You can help cmake by adding -DGeant4_PREFIX=<PATH>")
 endif()