+++ include/BDSCrystalFactory.hh
@@ -54,7 +54,8 @@ public:
 			    const BDSCrystalInfo* recipe);
 
 private:
-  /// Initialiser and can be used to reset factory pointers after use
+  /// Initialiser and can be used to reset factory pointers after use.
+  /// Note, non-virtual as used in constructor.
   void CleanUp();
   
   /// Common construction tasks.