+++ include/BDSGlobalConstants.hh
@@ -207,6 +207,8 @@ public:
   inline G4double StoreTrajectoryEnergyThreshold() const {return G4double (options.storeTrajectoryEnergyThreshold*CLHEP::GeV);}
   std::vector<int>                           StoreTrajectorySamplerIDs();
   std::vector<std::pair<double,double>>      StoreTrajectoryELossSRange();
+  inline G4bool   StoreSamplerAll()          const {return G4bool  (options.storeSamplerAll);}
+  inline G4bool   StoreSamplerRadius()       const {return G4bool  (options.storeSamplerRadius);}
   inline G4bool   StoreSamplerCharge()       const {return G4bool  (options.storeSamplerCharge);}
   inline G4bool   StoreSamplerKineticEnergy() const {return G4bool  (options.storeSamplerKineticEnergy);}
   inline G4bool   StoreSamplerMass()         const {return G4bool  (options.storeSamplerMass);}