+++ include/BDSOutput.hh
@@ -236,6 +236,7 @@ private:
   G4bool storeCollimationInfo;
   G4bool storeGeant4Data;
   G4bool storeModel;
+  G4bool storeSamplerRadius;
   G4bool storeSamplerCharge;
   G4bool storeSamplerKineticEnergy;
   G4bool storeSamplerMass;