+++ parser/optionsBase.cc
@@ -235,6 +235,8 @@ OptionsBase::OptionsBase()
   storeTrajectoryParticle   = "";
   storeTrajectoryParticleID = "";
   storeTrajectoryEnergyThreshold = -1.0;
+  storeSamplerAll          = false;
+  storeSamplerRadius       = false;
   storeSamplerCharge       = false;
   storeSamplerKineticEnergy = false;
   storeSamplerMass         = false;