+++ parser/optionsBase.h
@@ -267,6 +267,8 @@ namespace GMAD
     std::string storeTrajectoryParticleID;
     double      storeTrajectoryEnergyThreshold;
 
+    bool        storeSamplerAll;
+    bool        storeSamplerRadius;
     bool        storeSamplerCharge;
     bool        storeSamplerKineticEnergy;
     bool        storeSamplerMass;