+++ src/BDSIonDefinition.cc
@@ -75,7 +75,7 @@ void BDSIonDefinition::Parse(const G4String& definition)
 	  if (counter > 1)
 	    {// double
 	      G4double value = std::stod(match[1]);
-	      if (counter == 1)
+	      if (counter == 3)
 		{energy = value*CLHEP::keV;}
 	      else
 		{charge = value*CLHEP::eplus;}